pyproject.toml
[project]
name = "fmor-in"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "Pillow>=10.1",
    "pillow-heif>=0.18",
    "Jinja2>=3.1",
    "exifread>=3.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"
